+++ CloudShell.ResourceDefinitions.ResourceManager/ResourceModelResourceManagerAttributeProvider.cs
@@ -1,0 +1,8 @@
+using ResourceModelResource = CloudShell.ResourceDefinitions.Resource;
+
+namespace CloudShell.ResourceDefinitions.ResourceManager;
+
+public interface IResourceModelResourceManagerAttributeProvider
+{
+    IReadOnlyDictionary<string, string>? GetAttributes(ResourceModelResource resource);
+}