+++ CloudShell.ResourceDefinitions.ResourceManager/ResourceModelResourceManagerParentProvider.cs
@@ -1,0 +1,8 @@
+using ResourceModelResource = CloudShell.ResourceDefinitions.Resource;
+
+namespace CloudShell.ResourceDefinitions.ResourceManager;
+
+public interface IResourceModelResourceManagerParentProvider
+{
+    string? GetParentResourceId(ResourceModelResource resource);
+}